test+csv_to_database moved to test_csv_imported, minor changes
